Cleaned up the source, fixed MPI_Comm_create
Several collectives, such as MPI_Reduce are broken. They need
information about the communicator's size and/or root that is
not available with how things are currently implemented (write
event files, then reconstruct information about Comms and Groups).

Unfortunately, traces and in-situ simulator capture will require
two different implementations.
