+++ dumpi/libotf2dump/Makefile.am
@@ -13,7 +13,7 @@ AM_CPPFLAGS = -I$(top_srcdir) -I$(top_builddir)
 AM_LDFLAGS =
 library_includedir=$(includedir)/dumpi/libotf2dump
 library_include_HEADERS = \
-								      otf2writer.h
+                      otf2writer.h otf2writer-defs.h
 lib_LTLIBRARIES = libotf2dump.la
 libotf2dump_la_SOURCES = otf2writer.cc
 libotf2dump_la_LDFLAGS = -version-info @DUMPI_LIBVERSION@
